DAS unknown calls are now collected in a separate thread, once per second.
