+++ tw-entrypoints/src/main/java/com/transferwise/common/entrypoints/EntryPointsMetrics.java
@@ -17,7 +17,5 @@ public final class EntryPointsMetrics {
   public static final String TAG_RESOLUTION = "resolution";
   public static final String TAG_RESOLUTION_SUCCESS = "resolutionSuccess";
 
-  public static final String METRIC_PREFIX_ENTRYPOINTS = "EntryPoints.";
-
   public static final long MS_TO_NS = 1_000_000L;
 }
