+++ tw-entrypoints/src/main/java/com/transferwise/common/entrypoints/databaseaccessstatistics/DatabaseAccessStatisticsEntryPointInterceptor.java
@@ -6,7 +6,6 @@ import com.transferwise.common.context.TwContext;
 import com.transferwise.common.context.TwContextExecutionInterceptor;
 import com.transferwise.common.context.TwContextMetricsTemplate;
 import com.transferwise.common.entrypoints.EntryPointsMetrics;
-import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.DistributionSummary;
 import io.micrometer.core.instrument.Timer;
 import java.util.HashMap;
@@ -32,15 +31,6 @@ public class DatabaseAccessStatisticsEntryPointInterceptor implements TwContextE
   public static final String SUMMARY_REGISTERED_ROLLBACKS = "EntryPoints_Das_Registered_Rollbacks";
   public static final String SUMMARY_REGISTERED_COMMITS = "EntryPoints_Das_Registered_Commits";
 
-  public static final String COUNTER_UNREGISTERED_FETCHED_ROWS = "EntryPoints_Das_Unknown_FetchedRows";
-  public static final String COUNTER_UNREGISTERED_AFFECTED_ROWS = "EntryPoints_Das_Unknown_AffectedRows";
-  public static final String COUNTER_UNREGISTERED_EMPTY_TRANSACTIONS = "EntryPoints_Das_Unknown_EmptyTransactions";
-  public static final String COUNTER_UNREGISTERED_TIME_TAKEN_NS = "EntryPoints_Das_Unknown_TimeTakenNs";
-  public static final String COUNTER_UNREGISTERED_T_QUERIES = "EntryPoints_Das_Unknown_TQueries";
-  public static final String COUNTER_UNREGISTERED_NT_QUERIES = "EntryPoints_Das_Unknown_NTQueries";
-  public static final String COUNTER_UNREGISTERED_ROLLBACKS = "EntryPoints_Das_Unknown_Rollbacks";
-  public static final String COUNTER_UNREGISTERED_COMMITS = "EntryPoints_Das_Unknown_Commits";
-
   private final IMeterCache meterCache;
 
   public DatabaseAccessStatisticsEntryPointInterceptor(IMeterCache meterCache) {
@@ -57,7 +47,6 @@ public class DatabaseAccessStatisticsEntryPointInterceptor implements TwContextE
       return supplier.get();
     } finally {
       try {
-        registerUnknownCalls();
         registerCall(context, dbDasMap);
       } catch (Throwable t) {
         // TODO: Maybe should be throttled.
@@ -65,7 +54,6 @@ public class DatabaseAccessStatisticsEntryPointInterceptor implements TwContextE
       }
     }
   }
-
 
   private void registerCall(TwContext context, Map<String, DatabaseAccessStatistics> dbDasMap) {
     for (DatabaseAccessStatistics das : dbDasMap.values()) {
@@ -133,53 +121,6 @@ public class DatabaseAccessStatisticsEntryPointInterceptor implements TwContextE
     private Timer timeTakenNs;
   }
 
-  private void registerUnknownCalls() {
-    for (DatabaseAccessStatistics das : DatabaseAccessStatistics.unknownContextDbDasMap.values()) {
-      final long commits = das.getAndResetCommitsCount();
-      final long rollbacks = das.getAndResetRollbacksCount();
-      final long nonTransactionalQueries = das.getAndResetNonTransactionalQueriesCount();
-      final long transactionalQueries = das.getAndResetTransactionalQueriesCount();
-      final long timeTakenNs = das.getAndResetTimeTakenInDatabaseNs();
-      final long affectedRows = das.getAndResetAffectedRowsCount();
-      final long emptyTransactions = das.getAndResetEmptyTransactionsCount();
-      final long fetchedRows = das.getAndResetFetchedRowsCount();
-
-      TagsSet tagsSet = das.getTagsSet();
-      UnknownCallMeters meters = meterCache.metersContainer(METRIC_PREFIX_DAS + "unknownCallMetrics", tagsSet, () -> {
-        UnknownCallMeters result = new UnknownCallMeters();
-        result.commits = meterCache.counter(COUNTER_UNREGISTERED_COMMITS, tagsSet);
-        result.rollbacks = meterCache.counter(COUNTER_UNREGISTERED_ROLLBACKS, tagsSet);
-        result.nonTransactionalQueries = meterCache.counter(COUNTER_UNREGISTERED_NT_QUERIES, tagsSet);
-        result.transactionalQueries = meterCache.counter(COUNTER_UNREGISTERED_T_QUERIES, tagsSet);
-        result.timeTakenNs = meterCache.counter(COUNTER_UNREGISTERED_TIME_TAKEN_NS, tagsSet);
-        result.emptyTransactions = meterCache.counter(COUNTER_UNREGISTERED_EMPTY_TRANSACTIONS, tagsSet);
-        result.affectedRows = meterCache.counter(COUNTER_UNREGISTERED_AFFECTED_ROWS, tagsSet);
-        result.fetchedRows = meterCache.counter(COUNTER_UNREGISTERED_FETCHED_ROWS, tagsSet);
-        return result;
-      });
-
-      meters.commits.increment(commits);
-      meters.rollbacks.increment(rollbacks);
-      meters.nonTransactionalQueries.increment(nonTransactionalQueries);
-      meters.transactionalQueries.increment(transactionalQueries);
-      meters.timeTakenNs.increment(timeTakenNs);
-      meters.emptyTransactions.increment(emptyTransactions);
-      meters.affectedRows.increment(affectedRows);
-      meters.fetchedRows.increment(fetchedRows);
-    }
-  }
-
-  private static class UnknownCallMeters {
-
-    private Counter commits;
-    private Counter rollbacks;
-    private Counter nonTransactionalQueries;
-    private Counter transactionalQueries;
-    private Counter timeTakenNs;
-    private Counter emptyTransactions;
-    private Counter affectedRows;
-    private Counter fetchedRows;
-  }
 
   @Override
   public boolean applies(TwContext context) {
